requirements.txt
requests==2.31.0
colorama==0.4.6
python-dotenv==1.0.0
pytest==8.1.0
pytest-mock==3.12.0
